+++ jsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "@themost/oracle": ["index.js"]
+      "@themost/oracle": ["src/index"]
     }
   }
 }
