+++ package.json
@@ -2,9 +2,12 @@
   "name": "@themost/oracle",
   "version": "2.7.0",
   "description": "MOST Web Framework Oracle Data Adapter",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "module": "dist/index.esm.js",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "test": "jest"
+    "test": "jest",
+    "build": "rimraf dist && rollup --config rollup.config.js"
   },
   "engines": {
     "node": ">=8"
@@ -43,6 +46,8 @@
     "@babel/eslint-plugin": "^7.27.0",
     "@babel/plugin-proposal-decorators": "^7.25.9",
     "@babel/preset-env": "^7.26.9",
+    "@rollup/plugin-babel": "^5.3.1",
+    "@rollup/plugin-commonjs": "^22.0.0",
     "@themost/common": "^2.11.1",
     "@themost/data": "^2.18.5",
     "@themost/json-logger": "^1.3.0",
@@ -56,6 +61,9 @@
     "eslint-plugin-node": "^11.1.0",
     "jest": "^29.7.0",
     "mathjs": "^14.4.0",
-    "moment": "^2.30.1"
+    "moment": "^2.30.1",
+    "rollup": "^2.73.0",
+    "rollup-plugin-dts": "^4.2.1",
+    "typescript": "^4.6.4"
   }
 }
