+++ src/OracleAdapter.js
@@ -1,32 +1,13 @@
-// Copyright (c) 2017-2021, THEMOST LP. All rights reserved.
-const oracledb = require('oracledb');
-const async =  require('async');
-const util = require('util');
-const _ = require('lodash');
-const {SqlFormatter, SqlUtils, QueryField, QueryExpression} = require('@themost/query');
-const {TraceUtils,LangUtils}  = require('@themost/common');
+import oracledb from 'oracledb';
+import async from 'async';
+import util from 'util';
+import _ from 'lodash';
+import { SqlUtils, QueryExpression } from '@themost/query';
+import { TraceUtils, LangUtils } from '@themost/common';
+import { AsyncSeriesEventEmitter, before, after } from '@themost/events';
+import { OracleFormatter } from './OracleFormatter';
 
-
-function zeroPad(number, length) {
-    number = number || 0;
-    let res = number.toString();
-    while (res.length < length) {
-        res = '0' + res;
-    }
-    return res;
-}
-
-function instanceOf(any, ctor) {
-    // validate constructor
-    if (typeof ctor !== 'function') {
-        return false
-    }
-    // validate with instanceof
-    if (any instanceof ctor) {
-        return true;
-    }
-    return !!(any && any.constructor && any.constructor.name === ctor.name);
-}
+oracledb.fetchAsString = [ oracledb.CLOB ];
 
 /**
  *
@@ -43,6 +24,42 @@ function createLogger() {
     return newLogger;
 }
 
+/**
+ *
+ * @param {{target: OracleAdapter, query: string|QueryExpression, results: Array<*>}} event
+ */
+function onReceivingJsonObject(event) {
+    if (typeof event.query === 'object' && event.query.$select) {
+        // try to identify the usage of a $jsonObject dialect and format result as JSON
+        const { $select: select } = event.query;
+        if (select) {
+            const attrs = Object.keys(select).reduce((previous, current) => {
+                const fields = select[current];
+                previous.push(...fields);
+                return previous;
+            }, []).filter((x) => {
+                const [key] = Object.keys(x);
+                if (typeof key !== 'string') {
+                    return false;
+                }
+                return x[key].$jsonObject != null || x[key].$jsonArray != null  || x[key].$jsonGroupArray != null;
+            }).map((x) => {
+                return Object.keys(x)[0];
+            });
+            if (attrs.length > 0) {
+                if (Array.isArray(event.results)) {
+                    for(const result of event.results) {
+                        attrs.forEach((attr) => {
+                            if (Object.prototype.hasOwnProperty.call(result, attr) && typeof result[attr] === 'string') {
+                                result[attr] = JSON.parse(result[attr]);
+                            }
+                        });
+                    }
+                }
+            }
+        }
+    }
+}
 
 /**
  * @class
@@ -81,6 +98,9 @@ class OracleAdapter {
             }
         });
 
+        this.executing = new AsyncSeriesEventEmitter();
+        this.executed = new AsyncSeriesEventEmitter();
+        this.executed.subscribe(onReceivingJsonObject);
         /**
          * create a new instance of logger
          * @type {import('@themost/common').TraceLogger}
@@ -274,8 +294,16 @@ class OracleAdapter {
                 s = size > 0 ? util.format('NVARCHAR2(%s)', size) : 'NVARCHAR2(255)';
                 break;
             case 'Note':
-                s = size > 0 ? util.format('NVARCHAR2(%s)', size) : 'NVARCHAR2(2000)';
+                // important note: if size is greater than 4000 then we use CLOB instead of NVARCHAR2
+                if (size > 2000) {
+                    s = 'CLOB';
+                } else {
+                    s = size > 0 ? util.format('NVARCHAR2(%s)', size) : 'NVARCHAR2(2000)';
+                }
                 break;
+            case 'Json':
+                s = 'CLOB';
+                break
             case 'Image':
             case 'Binary':
                 s ='LONG RAW';
@@ -634,6 +662,18 @@ class OracleAdapter {
         });
     }
 
+    selectIdentityAsync(entity, attribute) {
+        const self = this;
+        return new Promise(function(resolve, reject) {
+            void self.selectIdentity(entity, attribute, function(err, value) {
+                if (err) {
+                    return reject(err);
+                }
+                return resolve(value);
+            });
+        });
+    }
+
     resetIdentity(entity, attribute, callback) {
         const self = this;
         return self.selectIdentity(entity, attribute , function(err) {
@@ -692,7 +732,7 @@ class OracleAdapter {
      * @param {*} batch
      * @param {function(Error=)} callback
      */
-    executeBatch(batch, callback) {
+    executeBatch(callback) {
         callback = callback || function() {};
         callback(new Error('DataAdapter.executeBatch() is obsolete. Use DataAdapter.executeInTransaction() instead.'));
     }
@@ -1294,6 +1334,30 @@ class OracleAdapter {
         return callback();
     }
 
+    @after(({target, args}, callback) => {
+        const [query, params] = args;
+        void target.executed.emit({
+            target,
+            query,
+            params
+        }).then(() => {
+            return callback();
+        }).catch((err) => {
+            return callback(err);
+        });
+    })
+    @before(({target, args}, callback) => {
+        const [query, params] = args;
+        void target.executing.emit({
+            target,
+            query,
+            params
+        }).then(() => {
+            return callback();
+        }).catch((err) => {
+            return callback(err);
+        });
+    })
     /**
      * Executes a query against the underlying database
      * @param query {QueryExpression|string|*}
@@ -1367,407 +1431,7 @@ class OracleAdapter {
     }
 }
 
-/**
- * @class
- * @augments {SqlFormatter}
- */
-class OracleFormatter extends SqlFormatter {
-    /**
-     * @constructor
-     */
-    constructor() {
-        super();
-        this.settings = {
-            nameFormat:OracleFormatter.NAME_FORMAT,
-            forceAlias:true,
-            useAliasKeyword: false
-        };
-    }
-
-    escapeName(name) {
-        if (typeof name === 'string')
-            return name.replace(/(\w+)/ig, this.settings.nameFormat);
-        return name;
-    }
-
-    /**
-     * Escapes an object or a value and returns the equivalent sql value.
-     * @param {*} value - A value that is going to be escaped for SQL statements
-     * @param {boolean=} unquoted - An optional value that indicates whether the resulted string will be quoted or not.
-     * returns {string} - The equivalent SQL string value
-     */
-    escape(value, unquoted) {
-        if (typeof value === 'boolean') { return value ? '1' : '0'; }
-        if (value instanceof Date) {
-            return util.format('TO_TIMESTAMP_TZ(%s, \'YYYY-MM-DD HH24:MI:SS.FF3TZH:TZM\')', this.escapeDate(value));
-        }
-        // if (typeof value === 'string' && LangUtils.isDate(value)) {
-        //     return util.format('TO_TIMESTAMP_TZ(%s, \'YYYY-MM-DD HH24:MI:SS.FF3TZH:TZM\')', this.escapeDate(new Date(value)));
-        // }
-        let res = super.escape.bind(this)(value, unquoted);
-        if (typeof value === 'string') {
-            if (/\\'/g.test(res)) {
-                //escape single quote (that is already escaped)
-                res = res.replace(/\\'/g, SINGLE_QUOTE_ESCAPE);
-                if (/\\"/g.test(res))
-                //escape double quote (that is already escaped)
-                    res = res.replace(/\\"/g, DOUBLE_QUOTE_ESCAPE);
-                if (/\\\\/g.test(res))
-                //escape slash (that is already escaped)
-                    res = res.replace(/\\\\/g, SLASH_ESCAPE);
-            }
-        }
-        return res;
-    }
-
-    /**
-     * @param {Date|*} val
-     * @returns {string}
-     */
-    escapeDate(val) {
-        const year   = val.getFullYear();
-        const month  = zeroPad(val.getMonth() + 1, 2);
-        const day    = zeroPad(val.getDate(), 2);
-        const hour   = zeroPad(val.getHours(), 2);
-        const minute = zeroPad(val.getMinutes(), 2);
-        const second = zeroPad(val.getSeconds(), 2);
-        //var millisecond = zeroPad(dt.getMilliseconds(), 3);
-        //format timezone
-        const offset = (new Date()).getTimezoneOffset(), timezone = (offset<=0 ? '+' : '-') + zeroPad(-Math.floor(offset/60),2) + ':' + zeroPad(offset%60,2);
-        return '\'' + year + '-' + month + '-' + day + ' ' + hour + ':' + minute + ':' + second + '.' + zeroPad(val.getMilliseconds(), 3) + timezone + '\'';
-    }
-
-
-    /**
-     * Formats a fixed query expression where select fields are constants e.g. SELECT 1 AS `id`,'John' AS `givenName` etc
-     * @param obj {QueryExpression|*}
-     * @returns {string}
-     */
-    formatFixedSelect(obj) {
-        let self = this;
-        let fields = obj.fields();
-        return 'SELECT ' + _.map(fields, function(x) { return self.format(x,'%f'); }).join(', ') + ' FROM DUAL';
-    }
-
-    /**
-     *
-     * @param {QueryExpression} obj
-     * @returns {string}
-     */
-    formatLimitSelect(obj) {
-
-        let sql;
-        const self=this;
-        let take = parseInt(obj.$take) || 0;
-        let skip = parseInt(obj.$skip) || 0;
-        if (take<=0) {
-            sql=self.formatSelect(obj);
-        }
-        else {
-            //add row_number with order
-            const keys = Object.keys(obj.$select);
-            if (keys.length === 0)
-                throw new Error('Entity is missing');
-            //get select fields
-            let selectFields = obj.$select[keys[0]];
-            //get order
-            let order = obj.$order;
-            //add row index field
-            selectFields.push({
-                '__RowIndex': {
-                  $row_index: order
-                }
-            });
-            //remove order
-            if (order) {
-                delete obj.$order;
-            }
-            //get sub query
-            const subQuery = self.formatSelect(obj);
-            //add order again
-            if (order) {
-                obj.$order = order;
-            }
-            //remove row index field
-            selectFields.pop();
-            const fields = [];
-            _.forEach(selectFields, (x) => {
-                if (typeof x === 'string') {
-                    fields.push(new QueryField(x));
-                }
-                else {
-                    /**
-                     * @type QueryField
-                     */
-                    let field = Object.assign(new QueryField(), x);
-                    fields.push(field.as() || field.getName());
-                }
-            });
-            sql = util.format('SELECT %s FROM (%s) t0 WHERE "__RowIndex" BETWEEN %s AND %s', _.map(fields, (x) => {
-                return self.format(x, '%f');
-            }).join(', '), subQuery, skip + 1, skip + take);
-        }
-        return sql;
-
-    }
-    isLogical(obj) {
-        let prop;
-        // eslint-disable-next-line no-unused-vars
-        for(let key in obj) {
-            if (Object.prototype.hasOwnProperty.call(obj, key)) {
-                prop = key;
-                break;
-            }
-        }
-        return (/^\$(and|or|not|nor)$/g.test(prop));
-    }
-    /**
-     * Implements [a & b] bitwise and expression formatter.
-     * @param p0 {*}
-     * @param p1 {*}
-     */
-    $bit(p0, p1)
-    {
-        return util.format('BITAND(%s, %s)', this.escape(p0), this.escape(p1));
-    }
-
-    /**
-     * Implements indexOf(str,substr) expression formatter.
-     * @param {string} p0 The source string
-     * @param {string} p1 The string to search for
-     * @returns {string}
-     */
-    $indexof(p0, p1) {
-        return util.format('(INSTR(%s,%s)-1)', this.escape(p0), this.escape(p1));
-    }
-
-    $indexOf(p0, p1) {
-        return util.format('(INSTR(%s,%s)-1)', this.escape(p0), this.escape(p1));
-    }
-
-    /**
-     * Implements contains(a,b) expression formatter.
-     * @param {string} p0 The source string
-     * @param {string} p1 The string to search for
-     * @returns {string}
-     */
-    $text(p0, p1) {
-        return util.format('(INSTR(%s,%s)-1)>=0', this.escape(p0), this.escape(p1));
-    }
-
-    /**
-     * Implements concat(a,b) expression formatter.
-     * @param {*} p0
-     * @param {*} p1
-     * @returns {string}
-     */
-    $concat(p0, p1) {
-        return util.format('CONCAT(%s,%s)', this.escape(p0),  this.escape(p1));
-    }
-
-    /**
-     * Implements substring(str,pos) expression formatter.
-     * @param {String} p0 The source string
-     * @param {Number} pos The starting position
-     * @param {Number=} length The length of the resulted string
-     * @returns {string}
-     */
-    $substring(p0, pos, length) {
-        if (length)
-            return util.format('SUBSTR(%s,%s,%s)', this.escape(p0), pos.valueOf()+1, length.valueOf());
-        else
-            return util.format('SUBSTR(%s,%s)', this.escape(p0), pos.valueOf()+1);
-    }
-
-    /**
-     * Implements length(a) expression formatter.
-     * @param {*} p0
-     * @returns {string}
-     */
-    $length(p0) {
-        return util.format('LENGTH(%s)', this.escape(p0));
-    }
-
-    /**
-     * @param {...*} p0
-     * @return {*}
-     */
-    // eslint-disable-next-line no-unused-vars
-    $row_index(p0) {
-        let args = Array.from(arguments).filter(function(x) {
-            return x != null;
-        });
-        return util.format('ROW_NUMBER() OVER(%s)', (args && args.length) ? this.format(args, '%o') : 'ORDER BY NULL');
-    }
-
-    $ceiling(p0) {
-        return util.format('CEIL(%s)', this.escape(p0));
-    }
-
-    $startswith(p0, p1) {
-        //validate params
-        if ( _.isNil(p0) ||  _.isNil(p1))
-            return '';
-        return 'REGEXP_COUNT(' + this.escape(p0) + ',\'^' + this.escape(p1, true) + '\', 1, \'i\')';
-    }
-
-    $contains(p0, p1) {
-        //validate params
-        if ( _.isNil(p0) ||  _.isNil(p1))
-            return '';
-        //(CASE WHEN REGEXP_COUNT(x, 'text', 1, 'i') > 0 THEN 1 ELSE 0 END)
-        return '(CASE WHEN REGEXP_COUNT(' + this.escape(p0) + ',\'' + this.escape(p1, true) + '\', 1, \'i\') > 0 THEN 1 ELSE 0 END)';
-    }
-
-    $endswith(p0, p1) {
-        //validate params
-        if ( _.isNil(p0) ||  _.isNil(p1))
-            return '';
-        return 'REGEXP_COUNT(' + this.escape(p0) + ',\'' + this.escape(p1, true) + '$\', 1, \'i\')';
-    }
-
-    $day(p0) {
-        return util.format('EXTRACT(DAY FROM %s)', this.escape(p0)) ;
-    }
-
-    $month(p0) {
-        return util.format('EXTRACT(MONTH FROM %s)', this.escape(p0)) ;
-    }
-
-    $year(p0) {
-        return util.format('EXTRACT(YEAR FROM %s)', this.escape(p0)) ;
-    }
-
-    $hour(p0) {
-        return util.format('EXTRACT(HOUR FROM %s)', this.escape(p0)) ;
-    }
-
-    $minute(p0) {
-        return util.format('EXTRACT(MINUTE FROM %s)', this.escape(p0)) ;
-    }
-
-    $second(p0) {
-        return util.format('EXTRACT(SECOND FROM %s)', this.escape(p0)) ;
-    }
-
-    $date(p0) {
-        //alternative date solution: 'TO_TIMESTAMP_TZ(TO_CHAR(%s, 'YYYY-MM-DD'),'YYYY-MM-DD')'
-        return util.format('TO_CHAR(TRUNC(%s), \'YYYY-MM-DD\')', this.escape(p0)) ;
-    }
-
-    /**
-     * Implements contains(a,b) expression formatter.
-     * @param {*} p0 The source string
-     * @param {string|*} p1 The string to search for
-     * @returns {string}
-     */
-    $regex(p0, p1) {
-        //validate params
-        if ( _.isNil(p0) ||  _.isNil(p1))
-            return '';
-        return 'REGEXP_LIKE(' + this.escape(p0) + ',\'' + this.escape(p1, true) + '\')';
-    }
-
-    /**
-     * @deprecated Use $ifNull() instead
-     * @param {*} p0 
-     * @param {*} p1 
-     * @returns 
-     */
-    $ifnull(p0, p1) {
-        return this.$ifNull(p0, p1) ;
-    }
-
-    $ifNull(p0, p1) {
-        return util.format('NVL(%s, %s)', this.escape(p0), this.escape(p1)) ;
-    }
-
-    $cond(ifExpr, thenExpr, elseExpr) {
-        // validate ifExpr which should an instance of QueryExpression or a comparison expression
-        let ifExpression;
-        if (instanceOf(ifExpr, QueryExpression)) {
-            ifExpression = this.formatWhere(ifExpr.$where);
-        } else if (this.isComparison(ifExpr) || this.isLogical(ifExpr)) {
-            ifExpression = this.formatWhere(ifExpr);
-        } else {
-            throw new Error('Condition parameter should be an instance of query or comparison expression');
-        }
-        return util.format('(CASE WHEN %s THEN %s ELSE %s END)', ifExpression, this.escape(thenExpr), this.escape(elseExpr));
-    }
-
-    $toString(p0) {
-        return util.format('TO_NCHAR(%s)', this.escape(p0)) ;
-    }
-
-    $uuid() {
-        return 'REGEXP_REPLACE(SYS_GUID(), \'(.{8})(.{4})(.{4})(.{4})(.{12})\', \'\\1-\\2-\\3-\\4-\\5\')';
-    }
-
-    /**
-     *
-     * @param {('date'|'datetime'|'timestamp')} type
-     * @returns
-     */
-    $getDate(type) {
-        switch (type) {
-            case 'date':
-                return 'TRUNC(SYSDATE)';
-            case 'datetime':
-                return 'SYSDATE';
-            case 'timestamp':
-                return 'CAST(SYSDATE AS TIMESTAMP WITH LOCAL TIME ZONE)';
-            default:
-                return 'CAST(SYSDATE AS TIMESTAMP WITH LOCAL TIME ZONE)';
-        }
-    }
-
-
-    $toInt(expr) {
-        return `CAST(${this.escape(expr)} AS INT)`;
-    }
-
-    $toDouble(expr) {
-        return this.$toDecimal(expr, 19, 8);
-    }
-
-    // noinspection JSCheckFunctionSignatures
-    /**
-     * @param {*} expr
-     * @param {number=} precision
-     * @param {number=} scale
-     * @returns
-     */
-    $toDecimal(expr, precision, scale) {
-        const p = typeof precision === 'number' ? Math.floor(precision) : 19;
-        const s = typeof scale === 'number' ? Math.floor(scale) : 8;
-        return `CAST(${this.escape(expr)} AS DECIMAL(${p},${s}))`;
-    }
-
-    $toLong(expr) {
-        return `CAST(${this.escape(expr)} AS NUMBER(19))`;
-    }
-
-
-
+export {
+    OracleAdapter
 }
 
-OracleFormatter.NAME_FORMAT = '"$1"';
-
-const SINGLE_QUOTE_ESCAPE = '\'\'';
-const DOUBLE_QUOTE_ESCAPE = '"';
-const SLASH_ESCAPE = '\\';
-
-/**
- * Creates an instance of OracleAdapter object that represents an Oracle database connection.
- * @param {*} options An object that represents the properties of the underlying database connection.
- * @returns {OracleAdapter}
- */
-function createInstance(options) {
-    return new OracleAdapter(options);
-}
-
-module.exports = {
-    OracleAdapter,
-    OracleFormatter,
-    createInstance
-}
