+++ src/OracleFormatter.d.ts
@@ -1,0 +1,4 @@
+import { SqlFormatter, QueryExpression } from "@themost/query";
+
+export declare class OracleFormatter extends SqlFormatter {
+}
