+++ src/index.d.ts
@@ -1,0 +1,3 @@
+export * from './OracleAdapter';
+export * from './OracleFormatter';
+export declare function createInstance(options: any): OracleAdapter;
