+++ src/index.js
@@ -1,0 +1,12 @@
+import { OracleFormatter } from './OracleFormatter';
+import { OracleAdapter } from './OracleAdapter';
+
+function createInstance(options) {
+    return new OracleAdapter(options);
+}
+
+export {
+    OracleAdapter,
+    OracleFormatter,
+    createInstance
+}
